stream_processor/app/requirements.txt
kafka-python
pandas
sqlalchemy
pyhive
sasl
thrift
thrift-sasl